+++ pages/api/v1/status/index.js
@@ -1,31 +1,14 @@
 import { createRouter } from "next-connect";
 import database from "infra/database.js";
-import { InternalServerError, MethodNotAllowedError } from "infra/errors";
+import controller from "infra/controller";
 
 const router = createRouter();
 
 router.get(getHandler);
 
-export default router.handler({
-  onNoMatch: onNoMatchHandler,
-  onError: onErrorHandler,
-});
-
-function onNoMatchHandler(request, response) {
-  const publicErrorObject = new MethodNotAllowedError();
-  response.status(publicErrorObject.statusCode).json(publicErrorObject);
-}
-
-function onErrorHandler(error, request, response) {
-  const publicErrorObject = new InternalServerError({
-    cause: error,
-  });
-  console.error("Error occurred while handling request:", publicErrorObject);
-  response.status(publicErrorObject.statusCode).json(publicErrorObject);
-}
+export default router.handler(controller.errorHandlers);
 
 // Handler principal para GET /api/v1/status. Consulta o status do banco e retorna em formato JSON.
-
 async function getHandler(_, response) {
   const updatedAt = new Date().toISOString();
   const appEnv = process.env.APP_ENV || process.env.NODE_ENV || "unknown";
